requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
certifi==2021.5.30
charset-normalizer==2.0.4
distlib==0.3.4
Django==3.2.6
djangorestframework==3.12.4
filelock==3.4.0
gunicorn==20.1.0
idna==3.2
numpy==1.21.1
pandas==1.3.1
Pillow==8.4.0
platformdirs==2.4.0
psycopg2==2.9.1
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
virtualenv==20.10.0
whitenoise==5.3.0
